+++ src/engine_api.rs
@@ -1357,9 +1357,9 @@ impl Engine {
 
         // Trims the JSON string and add a '#' in front
         let json_text = json.trim_start();
-        let scripts = if json_text.starts_with(Token::MapStart.keyword_syntax()) {
+        let scripts = if json_text.starts_with(Token::MapStart.literal_syntax()) {
             [json_text, ""]
-        } else if json_text.starts_with(Token::LeftBrace.keyword_syntax()) {
+        } else if json_text.starts_with(Token::LeftBrace.literal_syntax()) {
             ["#", json_text]
         } else {
             return Err(crate::ParseErrorType::MissingToken(
@@ -1741,40 +1741,36 @@ impl Engine {
         let lib = &[ast.lib()];
         self.eval_global_statements(scope, mods, &mut state, statements, lib, level)
     }
-    /// Evaluate a file, but throw away the result and only return error (if any).
-    /// Useful for when you don't need the result, but still need to keep track of possible errors.
+    /// Evaluate a file, returning any error (if any).
     ///
     /// Not available under `no_std` or `WASM`.
     #[cfg(not(feature = "no_std"))]
     #[cfg(not(any(target_arch = "wasm32", target_arch = "wasm64")))]
     #[inline(always)]
-    pub fn consume_file(&self, path: std::path::PathBuf) -> Result<(), Box<EvalAltResult>> {
-        Self::read_file(path).and_then(|contents| self.consume(&contents))
+    pub fn run_file(&self, path: std::path::PathBuf) -> Result<(), Box<EvalAltResult>> {
+        Self::read_file(path).and_then(|contents| self.run(&contents))
     }
-    /// Evaluate a file with own scope, but throw away the result and only return error (if any).
-    /// Useful for when you don't need the result, but still need to keep track of possible errors.
+    /// Evaluate a file with own scope, returning any error (if any).
     ///
     /// Not available under `no_std` or `WASM`.
     #[cfg(not(feature = "no_std"))]
     #[cfg(not(any(target_arch = "wasm32", target_arch = "wasm64")))]
     #[inline(always)]
-    pub fn consume_file_with_scope(
+    pub fn run_file_with_scope(
         &self,
         scope: &mut Scope,
         path: std::path::PathBuf,
     ) -> Result<(), Box<EvalAltResult>> {
-        Self::read_file(path).and_then(|contents| self.consume_with_scope(scope, &contents))
+        Self::read_file(path).and_then(|contents| self.run_with_scope(scope, &contents))
     }
-    /// Evaluate a string, but throw away the result and only return error (if any).
-    /// Useful for when you don't need the result, but still need to keep track of possible errors.
+    /// Evaluate a script, returning any error (if any).
     #[inline(always)]
-    pub fn consume(&self, script: &str) -> Result<(), Box<EvalAltResult>> {
-        self.consume_with_scope(&mut Default::default(), script)
+    pub fn run(&self, script: &str) -> Result<(), Box<EvalAltResult>> {
+        self.run_with_scope(&mut Default::default(), script)
     }
-    /// Evaluate a string with own scope, but throw away the result and only return error (if any).
-    /// Useful for when you don't need the result, but still need to keep track of possible errors.
+    /// Evaluate a script with own scope, returning any error (if any).
     #[inline]
-    pub fn consume_with_scope(
+    pub fn run_with_scope(
         &self,
         scope: &mut Scope,
         script: &str,
@@ -1790,18 +1786,16 @@ impl Engine {
             self.optimization_level,
         )?;
 
-        self.consume_ast_with_scope(scope, &ast)
+        self.run_ast_with_scope(scope, &ast)
     }
-    /// Evaluate an AST, but throw away the result and only return error (if any).
-    /// Useful for when you don't need the result, but still need to keep track of possible errors.
+    /// Evaluate an AST, returning any error (if any).
     #[inline(always)]
-    pub fn consume_ast(&self, ast: &AST) -> Result<(), Box<EvalAltResult>> {
-        self.consume_ast_with_scope(&mut Default::default(), ast)
+    pub fn run_ast(&self, ast: &AST) -> Result<(), Box<EvalAltResult>> {
+        self.run_ast_with_scope(&mut Default::default(), ast)
     }
-    /// Evaluate an [`AST`] with own scope, but throw away the result and only return error (if any).
-    /// Useful for when you don't need the result, but still need to keep track of possible errors.
+    /// Evaluate an [`AST`] with own scope, returning any error (if any).
     #[inline]
-    pub fn consume_ast_with_scope(
+    pub fn run_ast_with_scope(
         &self,
         scope: &mut Scope,
         ast: &AST,
@@ -1985,7 +1979,10 @@ impl Engine {
         let statements = ast.statements();
 
         if eval_ast && !statements.is_empty() {
+            // Make sure new variables introduced at global level do not _spill_ into the function call
+            let orig_scope_len = scope.len();
             self.eval_global_statements(scope, mods, state, statements, lib, 0)?;
+            scope.rewind(orig_scope_len);
         }
 
         let fn_def = ast
@@ -2151,8 +2148,8 @@ impl Engine {
     ///     }
     /// });
     ///
-    /// engine.consume("for x in range(0, 50000) {}")
-    ///     .expect_err("should error");
+    /// engine.run("for x in range(0, 50000) {}")
+    ///       .expect_err("should error");
     ///
     /// assert_eq!(*result.read().unwrap(), 9600);
     ///
@@ -2186,7 +2183,7 @@ impl Engine {
     /// let logger = result.clone();
     /// engine.on_print(move |s| logger.write().unwrap().push_str(s));
     ///
-    /// engine.consume("print(40 + 2);")?;
+    /// engine.run("print(40 + 2);")?;
     ///
     /// assert_eq!(*result.read().unwrap(), "42");
     /// # Ok(())
@@ -2219,7 +2216,7 @@ impl Engine {
     ///
     /// let mut ast = engine.compile(r#"let x = "hello"; debug(x);"#)?;
     /// ast.set_source("world");
-    /// engine.consume_ast(&ast)?;
+    /// engine.run_ast(&ast)?;
     ///
     /// #[cfg(not(feature = "no_position"))]
     /// assert_eq!(*result.read().unwrap(), r#"world @ 1:18 > "hello""#);