+++ tests/custom_syntax.rs
@@ -1,10 +1,12 @@
-use rhai::{Dynamic, Engine, EvalAltResult, LexError, ParseErrorType, Position, INT};
+use rhai::{
+    Dynamic, Engine, EvalAltResult, ImmutableString, LexError, ParseErrorType, Position, Scope, INT,
+};
 
 #[test]
 fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
     let mut engine = Engine::new();
 
-    engine.consume("while false {}")?;
+    engine.run("while false {}")?;
 
     // Disable 'while' and make sure it still works with custom syntax
     engine.disable_symbol("while");
@@ -19,21 +21,32 @@ fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
 
     engine.register_custom_syntax(
         &[
-            "exec", "[", "$ident$", ";", "$int$", "]", "->", "$block$", "while", "$expr$",
+            "exec", "[", "$ident$", "$symbol$", "$int$", "]", "->", "$block$", "while", "$expr$",
         ],
         true,
         |context, inputs| {
             let var_name = inputs[0].get_variable_name().unwrap().to_string();
-            let max = inputs[1].get_literal_value::<INT>().unwrap();
-            let stmt = &inputs[2];
-            let condition = &inputs[3];
+            let op = inputs[1].get_literal_value::<ImmutableString>().unwrap();
+            let max = inputs[2].get_literal_value::<INT>().unwrap();
+            let stmt = &inputs[3];
+            let condition = &inputs[4];
 
             context.scope_mut().push(var_name.clone(), 0 as INT);
 
             let mut count: INT = 0;
 
             loop {
-                if count >= max {
+                let done = match op.as_str() {
+                    "<" => count >= max,
+                    "<=" => count > max,
+                    ">" => count <= max,
+                    ">=" => count < max,
+                    "==" => count != max,
+                    "!=" => count == max,
+                    _ => return Err(format!("Unsupported operator: {}", op).into()),
+                };
+
+                if done {
                     break;
                 }
 
@@ -64,11 +77,18 @@ fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
         },
     )?;
 
+    assert!(matches!(
+        *engine
+            .run("let foo = (exec [x<<15] -> { x += 2 } while x < 42) * 10;")
+            .expect_err("should error"),
+        EvalAltResult::ErrorRuntime(_, _)
+    ));
+
     assert_eq!(
         engine.eval::<INT>(
             "
                 let x = 0;
-                let foo = (exec [x;15] -> { x += 2 } while x < 42) * 10;
+                let foo = (exec [x<15] -> { x += 2 } while x < 42) * 10;
                 foo
             "
         )?,
@@ -78,7 +98,7 @@ fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
         engine.eval::<INT>(
             "
                 let x = 0;
-                exec [x;100] -> { x += 1 } while x < 42;
+                exec [x<100] -> { x += 1 } while x < 42;
                 x
             "
         )?,
@@ -87,7 +107,7 @@ fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
     assert_eq!(
         engine.eval::<INT>(
             "
-                exec [x;100] -> { x += 1 } while x < 42;
+                exec [x<100] -> { x += 1 } while x < 42;
                 x
             "
         )?,
@@ -97,7 +117,7 @@ fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
         engine.eval::<INT>(
             "
                 let foo = 123;
-                exec [x;15] -> { x += 1 } while x < 42;
+                exec [x<15] -> { x += 1 } while x < 42;
                 foo + x + x1 + x2 + x3
             "
         )?,
@@ -115,6 +135,51 @@ fn test_custom_syntax() -> Result<(), Box<EvalAltResult>> {
             "Improper symbol for custom syntax at position #1: '!'".to_string()
         ))
     );
+
+    // Check self-termination
+    engine
+        .register_custom_syntax(&["test1", "$block$"], true, |_, _| Ok(Dynamic::UNIT))?
+        .register_custom_syntax(&["test2", "}"], true, |_, _| Ok(Dynamic::UNIT))?
+        .register_custom_syntax(&["test3", ";"], true, |_, _| Ok(Dynamic::UNIT))?;
+
+    assert_eq!(engine.eval::<INT>("test1 { x = y + z; } 42")?, 42);
+    assert_eq!(engine.eval::<INT>("test2 } 42")?, 42);
+    assert_eq!(engine.eval::<INT>("test3; 42")?, 42);
+
+    // Register the custom syntax: var x = ???
+    engine.register_custom_syntax(
+        &["var", "$ident$", "=", "$expr$"],
+        true,
+        |context, inputs| {
+            let var_name = inputs[0].get_variable_name().unwrap().to_string();
+            let expr = &inputs[1];
+
+            // Evaluate the expression
+            let value = context.eval_expression_tree(expr)?;
+
+            if !context.scope().is_constant(&var_name).unwrap_or(false) {
+                context.scope_mut().set_value(var_name, value);
+                Ok(Dynamic::UNIT)
+            } else {
+                Err(format!("variable {} is constant", var_name).into())
+            }
+        },
+    )?;
+
+    let mut scope = Scope::new();
+
+    assert_eq!(
+        engine.eval_with_scope::<INT>(&mut scope, "var foo = 42; foo")?,
+        42
+    );
+    assert_eq!(scope.get_value::<INT>("foo"), Some(42));
+    assert_eq!(scope.len(), 1);
+    assert_eq!(
+        engine.eval_with_scope::<INT>(&mut scope, "var foo = 123; foo")?,
+        123
+    );
+    assert_eq!(scope.get_value::<INT>("foo"), Some(123));
+    assert_eq!(scope.len(), 1);
 
     Ok(())
 }