+++ tests/while_loop.rs
@@ -57,9 +57,9 @@ fn test_infinite_loops() -> Result<(), Box<EvalAltResult>> {
 
     engine.set_max_operations(1024);
 
-    assert!(engine.consume("loop {}").is_err());
-    assert!(engine.consume("while true {}").is_err());
-    assert!(engine.consume("do {} while true").is_err());
+    assert!(engine.run("loop {}").is_err());
+    assert!(engine.run("while true {}").is_err());
+    assert!(engine.run("do {} while true").is_err());
 
     Ok(())
 }